requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.57.0
accelerate>=0.25.0

# Qwen-VL specific
qwen-vl-utils>=0.0.8

# ModelScope support (国内镜像)
modelscope>=1.9.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
httpx>=0.26.0
aiofiles>=23.2.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
